Add files via upload
Added version 2

Improved handling of detector geometry and physics values through new Detector structs

Now have four electron LL hypothesis models. One for 1 MIP energy hypothesis, one for helical track hypothesis, one for max hits in tracker hypothesis, one for ToF being in agreement with speed of light hypothesis.

In addition to this, rewrote existing code to use the new detector structs and data loader struct. 

Created a LogLikelihood model making code that outputs the results for the LL values to a csv so that one can use it later to compute probability that a given event satisfies certain hypotheses and is similar to the model particle. Currently including values for electron.